many small changes and implemented full adc code to temperature conversion

+++ src/telemetry.rs
@@ -48,14 +48,14 @@ pub struct TelemetryClient<T: Serialize> {
 /// overhead.
 #[derive(Serialize)]
 pub struct Telemetry {
-    pub adc: [u32; 2],
+    pub adc: [f64; 2],
     pub dac: [u32; 2],
 }
 
 impl Default for Telemetry {
     fn default() -> Self {
         Self {
-            adc: [0, 0],
+            adc: [0.0, 0.0],
             dac: [0x20000, 0x20000],
         }
     }
